--- server/routes/index.js
@@ -1,7 +1,0 @@
-// router.get('/test', async (ctx, next) => {
-//     ctx.body={test:true};
-// });
-
-
-
-// module.export=router;
